Move IngestBagRequestGenerators into the common package
