+++ archive/common/src/test/scala/uk/ac/wellcome/platform/archive/common/generators/IngestBagRequestGenerators.scala
@@ -1,18 +1,14 @@
-package uk.ac.wellcome.platform.archive.archivist.models
+package uk.ac.wellcome.platform.archive.common.generators
 
 import java.net.URI
 import java.util.UUID
 
 import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
-import uk.ac.wellcome.platform.archive.common.models.{
-  IngestBagRequest,
-  StorageSpace
-}
+import uk.ac.wellcome.platform.archive.common.models.{IngestBagRequest, StorageSpace}
 import uk.ac.wellcome.storage.ObjectLocation
 
-trait IngestRequestContextGenerators extends RandomThings {
-
-  def createIngestBagRequest = createIngestBagRequestWith()
+trait IngestBagRequestGenerators extends RandomThings {
+  def createIngestBagRequest: IngestBagRequest = createIngestBagRequestWith()
 
   def createIngestBagRequestWith(requestId: UUID = randomUUID,
                                  ingestBagLocation: ObjectLocation =
